requirements.txt
fastapi
uvicorn
python-multipart
requests
pytest
python-dotenv
